requirements.txt
google-generativeai
instagrapi
playwright
python-dotenv
schedule
requests
